backend/requirements.txt
# Framework Web
fastapi==0.115.6
uvicorn==0.34.0

# Base de données
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.1

# Sécurité et authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20

# HTTP client pour l'API météo
httpx==0.28.1
requests==2.32.3

# Variables d'environnement
python-dotenv==1.0.1

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Tests
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0